backend/requirements.txt
# backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers>=4.0.0
Pillow==10.4.0
python-decouple==3.8
requests>=2.31.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# For production deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Production database (uncomment when needed)
# psycopg2-binary==2.9.9

# Development tools (optional)
django-debug-toolbar==4.2.0
pytest-django==4.8.0
factory-boy==3.3.0